requirements.txt
python-Levenshtein==0.10.2
wsgiref==0.1.2
